[INJIMOB-2810] move constants to constants folder
