+++ Sources/OpenID4VP/NetworkManager/NetworkManager.swift
@@ -54,22 +54,3 @@ public struct NetworkManager: NetworkManaging {
         }
     }
 }
-
-
-public enum HTTP_METHOD: String {
-    case POST
-    case GET
-}
-
-public typealias HttpMethod = HTTPMethod
-
-enum Header : String {
-    case contentType = "Content-Type"
-}
-
-
-public enum ContentTypes : String {
-    case applicationJson = "application/json"
-    case applicationJwt = "application/oauth-authz-req+jwt"
-    case applicationFormUrlEncoded = "application/x-www-form-urlencoded"
-}
